Added a functional BP_RollingFlower
Each BP_RollingFlower has instance editable unroll length, unroll time (s), and extension time (s).

As far as I know, it works well.  You just can't scale it on the X axis in the level as it will not extend correctly.  The code bases it's extension off of the X scale.  I'm sure there's a way around this.